requirements.txt
annotated-types==0.5.0
anyio==3.7.1
astroid==2.15.6
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dill==0.3.7
distlib==0.3.7
fastapi==0.103.1
filelock==3.12.2
h11==0.14.0
httptools==0.6.0
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
pipenv==2023.8.25
platformdirs==3.10.0
psycopg2-binary==2.9.7
pydantic==2.3.0
pydantic_core==2.6.3
pylint==2.17.5
python-dotenv==1.0.0
pywatchman==1.4.1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
tomlkit==0.12.1
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
virtualenv==20.24.3
watchfiles==0.20.0
websockets==11.0.3
wrapt==1.15.0
